👌 IMPROVE: Create common comment component
